fusion + temp_array: widen offset math and guard negative size
simulate_fusion_at_array.cpp: all six fused ArrayAt macros computed the
element address as `rr*stride + offset` in uint32, mirroring the same
bug already fixed in SimNode_ArrayAt / SafeArrayAt. Now that arrays can
hold >4GB of data, rr*stride can wrap to a smaller pointer even when
rr < pl->size. Widened to uint64 multiply at every site.

module_builtin_runtime.cpp: the int version of builtin_make_temp_array
wrapped negative sizes through uint32 into huge uint64 capacity,
mirroring the bug we already fixed in the i64 overload. Clamp to empty
the same way.

+++ src/simulate/simulate_fusion_at_array.cpp
@@ -44,7 +44,7 @@ namespace das {
             auto pl = (Array *) l.compute##COMPUTEL(context); \
             auto rr = uint32_t(r.subexpr->evalInt(context)); \
             if ( rr >= pl->size ) context.throw_error_at(debugInfo,"array index out of range, %u of %llu", rr, (unsigned long long)pl->size); \
-            return *((CTYPE *)(pl->data + rr*stride + offset)); \
+            return *((CTYPE *)(pl->data + uint64_t(rr)*uint64_t(stride) + offset)); \
         } \
         DAS_NODE(TYPE,CTYPE); \
     };
@@ -56,7 +56,7 @@ namespace das {
             auto pl = (Array *) l.compute##COMPUTEL(context); \
             auto rr = *((uint32_t *)r.compute##COMPUTER(context)); \
             if ( rr >= pl->size ) context.throw_error_at(debugInfo,"array index out of range, %u of %llu", rr, (unsigned long long)pl->size); \
-            return *((CTYPE *)(pl->data + rr*stride + offset)); \
+            return *((CTYPE *)(pl->data + uint64_t(rr)*uint64_t(stride) + offset)); \
         } \
         DAS_NODE(TYPE,CTYPE); \
     };
@@ -86,7 +86,7 @@ namespace das {
             auto rr = uint32_t(r.subexpr->evalInt(context)); \
             if ( rr >= pl->size ) context.throw_error_at(debugInfo,"array index out of range, %u of %llu", rr, (unsigned long long)pl->size); \
             vec4f __r; \
-            DAS_LDU_WORKHORSE(__r, pl->data + rr*stride + offset, CTYPE); \
+            DAS_LDU_WORKHORSE(__r, pl->data + uint64_t(rr)*uint64_t(stride) + offset, CTYPE); \
             return __r; \
         } \
     };
@@ -100,7 +100,7 @@ namespace das {
             auto rr = *((uint32_t *)r.compute##COMPUTER(context)); \
             if ( rr >= pl->size ) context.throw_error_at(debugInfo,"array index out of range, %u of %llu", rr, (unsigned long long)pl->size); \
             vec4f __r; \
-            DAS_LDU_WORKHORSE(__r, pl->data + rr*stride + offset, CTYPE); \
+            DAS_LDU_WORKHORSE(__r, pl->data + uint64_t(rr)*uint64_t(stride) + offset, CTYPE); \
             return __r; \
         } \
     };
@@ -120,7 +120,7 @@ namespace das {
             auto pl = (Array *) l.compute##COMPUTEL(context); \
             auto rr = uint32_t(r.subexpr->evalInt(context)); \
             if ( rr >= pl->size ) context.throw_error_at(debugInfo,"array index out of range, %u of %llu", rr, (unsigned long long)pl->size); \
-            return pl->data + rr*stride + offset; \
+            return pl->data + uint64_t(rr)*uint64_t(stride) + offset; \
         } \
         DAS_PTR_NODE; \
     };
@@ -133,7 +133,7 @@ namespace das {
             auto pl = (Array *) l.compute##COMPUTEL(context); \
             auto rr = *((uint32_t *)r.compute##COMPUTER(context)); \
             if ( rr >= pl->size ) context.throw_error_at(debugInfo,"array index out of range, %u of %llu", rr, (unsigned long long)pl->size); \
-            return pl->data + rr*stride + offset; \
+            return pl->data + uint64_t(rr)*uint64_t(stride) + offset; \
         } \
         DAS_PTR_NODE; \
     };
